fix(compiler-types): move shared compiler constants
