+++ packages/compiler-types/src/diagnostics.ts
@@ -1,0 +1,37 @@
+import type { CompilationContext, InstructionCompiler } from './semantic';
+
+/**
+ * Internal compiler-stage error shape returned by getError().
+ * This is not the public cross-stage contract; consumers should use CompilerDiagnostic.
+ */
+export interface CompilerStageError {
+	message: string;
+	line: Parameters<InstructionCompiler>[0];
+	context?: CompilationContext;
+	code: number;
+}
+
+/**
+ * The shared, serializable diagnostic shape exposed to all consumers of the compiler pipeline.
+ * Both syntax errors (SyntaxRulesError) and semantic/compiler errors conform to this contract
+ * once serialized. Consumers must not special-case either stage.
+ */
+export interface CompilerDiagnosticLine {
+	lineNumberBeforeMacroExpansion: number;
+	lineNumberAfterMacroExpansion: number;
+	instruction?: string;
+	arguments?: unknown[];
+}
+
+export interface CompilerDiagnosticContext {
+	codeBlockId?: string;
+	codeBlockType?: 'module' | 'function' | 'constants';
+}
+
+export interface CompilerDiagnostic {
+	/** Numeric ErrorCode for compiler errors; SyntaxErrorCode string for syntax errors. */
+	code: number | string;
+	message: string;
+	line: CompilerDiagnosticLine;
+	context: CompilerDiagnosticContext;
+}
