+++ packages/compiler-types/src/instructions.ts
@@ -1,0 +1,76 @@
+export type Instruction =
+	| 'and'
+	| 'or'
+	| 'load'
+	| 'load8u'
+	| 'load16u'
+	| 'load8s'
+	| 'load16s'
+	| 'clampAddress'
+	| 'clampModuleAddress'
+	| 'clampGlobalAddress'
+	| 'localSet'
+	| 'else'
+	| 'if'
+	| 'ifEnd'
+	| 'lessThan'
+	| 'store'
+	| 'sub'
+	| 'div'
+	| 'xor'
+	| 'local'
+	| 'greaterOrEqual'
+	| 'add'
+	| 'min'
+	| 'max'
+	| 'greaterThan'
+	| 'branch'
+	| 'branchIfTrue'
+	| 'exitIfTrue'
+	| 'push'
+	| 'block'
+	| 'blockEnd'
+	| 'lessOrEqual'
+	| 'mul'
+	| 'loop'
+	| 'loopIndex'
+	| 'loopEnd'
+	| 'greaterOrEqualUnsigned'
+	| 'equalToZero'
+	| 'notEqual'
+	| 'notZero'
+	| 'shiftLeft'
+	| 'shiftRight'
+	| 'shiftRightUnsigned'
+	| 'remainder'
+	| 'castToInt'
+	| 'castToFloat'
+	| 'castToFloat64'
+	| 'drop'
+	| 'clearStack'
+	| 'risingEdge'
+	| 'fallingEdge'
+	| 'hasChanged'
+	| 'abs'
+	| 'equal'
+	| 'branchIfUnchanged'
+	| 'sqrt'
+	| 'loadFloat'
+	| 'round'
+	| 'ensureNonZero'
+	| 'function'
+	| 'functionEnd'
+	| 'return'
+	| 'param'
+	| 'call'
+	| '#skipExecution'
+	| '#initOnly'
+	| '#impure'
+	| '#export'
+	| '#loopCap'
+	| 'mapBegin'
+	| 'map'
+	| 'default'
+	| 'mapEnd'
+	| 'storeBytes'
+	| 'memoryCopy';
