+++ packages/compiler-types/src/macros.ts
@@ -1,0 +1,11 @@
+export interface ExpandedLine {
+	line: string;
+	callSiteLineNumber: number;
+	macroId?: string;
+}
+
+export interface MacroDefinition {
+	name: string;
+	body: string[];
+	definitionLineNumber: number;
+}
