+++ packages/compiler-types/src/runtime.ts
@@ -1,0 +1,32 @@
+import type { ASTCacheStats } from './ast';
+import type { CompiledFunctionLookup, CompiledModuleLookup } from './compiled';
+
+export type MemoryReinitReason =
+	| { kind: 'no-instance' }
+	| { kind: 'memory-size-changed'; prevBytes: number; nextBytes: number }
+	| { kind: 'memory-structure-changed' };
+
+export type MemoryAction = { action: 'reused' } | { action: 'recreated'; reason: MemoryReinitReason };
+
+export type GetOrCreateMemoryResult = {
+	memoryRef: WebAssembly.Memory;
+	memoryAction: MemoryAction;
+};
+
+export type GetOrCreateWasmInstanceResult = {
+	wasmInstanceRef: WebAssembly.Instance;
+	hasWasmInstanceBeenReset: boolean;
+};
+
+export type CompileAndUpdateMemoryResult = {
+	codeBuffer: Uint8Array;
+	compiledModules: CompiledModuleLookup;
+	compiledFunctions?: CompiledFunctionLookup;
+	requiredMemoryBytes: number;
+	allocatedMemoryBytes: number;
+	astCacheStats: ASTCacheStats;
+	memoryRef: WebAssembly.Memory;
+	hasWasmInstanceBeenReset: boolean;
+	memoryAction: MemoryAction;
+	initOnlyReran: boolean;
+};
