+++ packages/compiler/src/compilerError.ts
@@ -17,7 +17,7 @@
  *   before any semantic context is built → use SyntaxRulesError in syntaxError.ts.
  */
 
-import { SUPPORTED_MEMORY_ACCESS_BYTE_WIDTHS } from './consts';
+import { SUPPORTED_MEMORY_ACCESS_BYTE_WIDTHS } from '@8f4e/compiler-types';
 
 import type { AST, CompilationContext, CompilerStageError } from '@8f4e/compiler-types';
