+++ packages/compiler/src/index.ts
@@ -19,6 +19,7 @@ import {
 	createPassiveDataSegment,
 	WASM_MEMORY_PAGE_SIZE,
 } from '@8f4e/compiler-wasm-utils';
+import { GLOBAL_ALIGNMENT_BOUNDARY } from '@8f4e/compiler-types';
 
 import { compileModule, compileFunction } from './compiler';
 import createBufferFunctionBody from './wasmBuilders/createBufferFunctionBody';
@@ -28,7 +29,7 @@ import {
 	collectNamespacesFromASTs,
 	collectFunctionMetadataFromAsts,
 } from './semantic/buildNamespace';
-import { EXPORTED_FUNCTION_COUNT, GLOBAL_ALIGNMENT_BOUNDARY, HEADER, VERSION } from './consts';
+import { EXPORTED_FUNCTION_COUNT, HEADER, VERSION } from './consts';
 import sortModules from './graphOptimizer';
 import { createInitialMemoryDataSegments } from './initialMemoryDataSegments';
 import { ErrorCode, getError } from './compilerError';
@@ -55,7 +56,7 @@ export {
 	HALF_WORD_MEMORY_ACCESS_WIDTH,
 	WORD_MEMORY_ACCESS_WIDTH,
 	DOUBLE_WORD_MEMORY_ACCESS_WIDTH,
-} from './consts';
+} from '@8f4e/compiler-types';
 export { default as instructions } from './instructionCompilers';
 export {
 	prepassNamespace,
