+++ packages/compiler/src/semantic/declarations/array.ts
@@ -1,12 +1,12 @@
 import {
 	ArgumentType,
+	GLOBAL_ALIGNMENT_BOUNDARY,
 	type ArrayDeclarationLine,
 	type CompilationContext,
 	type InstructionCompiler,
 	type MemoryTypes,
 } from '@8f4e/compiler-types';
 
-import { GLOBAL_ALIGNMENT_BOUNDARY } from '../../consts';
 import { ErrorCode, getError } from '../../compilerError';
 import { alignAbsoluteWordOffset, getAbsoluteWordOffset, getByteAddressFromWordOffset } from '../layoutAddresses';
