+++ packages/compiler/src/semantic/layoutAddresses.ts
@@ -1,4 +1,4 @@
-import { GLOBAL_ALIGNMENT_BOUNDARY } from '../consts';
+import { GLOBAL_ALIGNMENT_BOUNDARY } from '@8f4e/compiler-types';
 
 export function getByteAddressFromWordOffset(startingByteAddress: number, wordOffset: number): number {
 	return startingByteAddress + wordOffset * GLOBAL_ALIGNMENT_BOUNDARY;
