+++ packages/compiler/src/semantic/resolveCompileTimeArgument.ts
@@ -1,8 +1,8 @@
 import { ArgumentType } from '@8f4e/tokenizer';
+import { GLOBAL_ALIGNMENT_BOUNDARY } from '@8f4e/compiler-types';
 
 import { getEndByteAddress, getModuleEndByteAddress } from './layoutAddresses';
 
-import { GLOBAL_ALIGNMENT_BOUNDARY } from '../consts';
 import {
 	getElementCount,
 	getElementMaxValue,
