+++ packages/compiler/src/utils/internalResources.ts
@@ -1,4 +1,4 @@
-import { GLOBAL_ALIGNMENT_BOUNDARY } from '../consts';
+import { GLOBAL_ALIGNMENT_BOUNDARY } from '@8f4e/compiler-types';
 
 import type { CompilationContext, InternalResource } from '@8f4e/compiler-types';
